+++ scratch.js
@@ -10,10 +10,7 @@ const fromFile = require('./tests/lib/_cachedPage');
 //doc.infoboxes('Venue')
 
 let doc = fromFile('royal_cinema');
-console.log(doc.sentences(0).json({
-  links: false,
-  formatting: false
-}));
+console.log(doc.section(0).json());
 
 // wtf.fetch('Royal Cinema').then((doc) => {
 //   console.log(doc.json({